Add acceptance tests step to CI
We have updated the CI shim such that it will automatically
create the $HOME/.kube/config file if/when the $KUBECONFIG_FILE_CONTENTS
env variable is set. This isn't great ... but bypasses the need to check
credentials into a GitHub repo somewhere.

We are making the assumption for now that anyone using sm will need to
have a kubeconfig file at $HOME/.kube/config.

Also, we are assuming a basic auth user (for now) for connecting to
k8s.

[#163694822]

Co-authored-by: William Martin <wmartin@pivotal.io>